[WIP] More aggressive clippy
Use newtype for `ScorerMatrix`.
